Avans-Web/src/app/callback/callback.component.ts
import { Component, OnInit } from '@angular/core';
import {AuthService} from '../services/auth.service';
import {ActivatedRoute, Router} from '@angular/router';

@Component({
  selector: 'app-callback',
  templateUrl: './callback.component.html',
  styleUrls: ['./callback.component.scss']
})
export class CallbackComponent implements OnInit {

  constructor(
    private authService: AuthService,
    private router: Router,
    private route: ActivatedRoute
  ) { }

  ngOnInit(): void {
    this.route.queryParamMap.subscribe(params => {
      if (params.get('access_token')) {
        this.authService.setAccessToken(params.get('access_token'));
        this.router.navigate(['/admin-dashboard']);
      }
    });
  }

}
